requirements.txt
exifread==3.5.1
folium==0.20.0
keras==3.12.0
matplotlib==3.10.8
numpy==2.3.5
Pillow==12.0.0
scikit_learn==1.8.0
seaborn==0.13.2
tensorflow==2.20.0
